# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mini-stim-server-soma",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/bucket/{soul_id}/{session_id}/{key}": {
      "get": {
        "tags": [
          "bucket"
        ],
        "operationId": "get_bucket_object",
        "parameters": [
          {
            "name": "soul_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bucket object bytes",
            "content": {
              "application/octet-stream": {}
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "tags": [],
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "create_session",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "tags": [],
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [],
        "operationId": "update_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSummary"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/materials": {
      "post": {
        "tags": [],
        "operationId": "session_material",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionMaterial"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/messages": {
      "get": {
        "tags": [],
        "operationId": "list_messages",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionMessage"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/runtime": {
      "get": {
        "tags": [],
        "operationId": "runtime_snapshot",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRuntimeSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/send": {
      "post": {
        "tags": [],
        "operationId": "send_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendSessionAcceptedResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Compact": {
        "type": "object",
        "required": [
          "id",
          "turn_id",
          "summary",
          "start_session_seq",
          "end_session_seq",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "end_session_seq": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "start_session_seq": {
            "type": "integer",
            "format": "int64"
          },
          "summary": {
            "type": "string"
          },
          "turn_id": {
            "type": "string"
          }
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "required": [
          "session"
        ],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "MaterialKind": {
        "type": "string",
        "enum": [
          "system_prompt"
        ]
      },
      "MaterialRequest": {
        "type": "object",
        "required": [
          "kind"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/MaterialKind"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "id",
          "actor_type",
          "actor_id",
          "content",
          "state",
          "version",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "actor_type": {
            "$ref": "#/components/schemas/ActorType"
          },
          "content": {
            "$ref": "#/components/schemas/MessageContent"
          },
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "deleted_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/String"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/MessageState"
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "MessageState": {
        "type": "string",
        "enum": [
          "pending",
          "fixed"
        ]
      },
      "SendSessionAcceptedResponse": {
        "type": "object",
        "required": [
          "session",
          "soul_session",
          "soul_profile",
          "turn",
          "user_message"
        ],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          },
          "soul_profile": {
            "$ref": "#/components/schemas/SoulProfile"
          },
          "soul_session": {
            "$ref": "#/components/schemas/SoulSession"
          },
          "turn": {
            "$ref": "#/components/schemas/Turn"
          },
          "user_message": {
            "$ref": "#/components/schemas/SessionMessage"
          }
        }
      },
      "SendSessionRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessagePart"
            }
          }
        }
      },
      "Session": {
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "fork_point": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "parent_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          }
        }
      },
      "SessionDetail": {
        "type": "object",
        "required": [
          "session",
          "profile",
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionMessage"
            }
          },
          "profile": {
            "$ref": "#/components/schemas/SessionProfile"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
...
      "SessionMaterial": {
        "type": "object",
        "required": [
          "session_id",
          "kind",
          "content_type",
          "text",
          "updated_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/MaterialKind"
          },
          "session_id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          }
        }
      },
      "SessionMessage": {
        "type": "object",
        "required": [
          "relation",
          "message",
          "content_text"
        ],
        "properties": {
          "content_text": {
            "type": "string"
          },
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "relation": {
            "$ref": "#/components/schemas/SessionMessageRef"
          }
        }
      },
      "SessionMessageRef": {
        "type": "object",
        "required": [
          "session_id",
          "message_id",
          "session_seq",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "message_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "session_seq": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SessionProfile": {
        "type": "object",
        "required": [
          "session_id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "desc": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_id": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          }
        }
      },
      "SessionRuntimeSnapshot": {
        "type": "object",
        "required": [
          "session",
          "profile",
          "messages",
          "turns",
          "thinking_spans",
          "tool_calls",
          "tool_results",
          "compacts",
          "effects"
        ],
        "properties": {
          "compacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Compact"
            }
          },
          "effects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionEffect"
            }
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionMessage"
            }
          },
          "profile": {
            "$ref": "#/components/schemas/SessionProfile"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          },
          "soul_profile": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SoulProfile"
              }
            ]
          },
          "soul_session": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SoulSession"
              }
            ]
          },
          "thinking_spans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThinkingSpan"
            }
          },
          "tool_calls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            }
          },
          "tool_results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolResult"
            }
          },
          "turns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Turn"
            }
          }
        }
      },
      "SessionSummary": {
        "type": "object",
        "required": [
          "session",
          "profile"
        ],
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/SessionProfile"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
...
      "String": {
        "type": "string"
      },
...
      "UpdateSessionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  }
}